Time: 1 ms (99.81%), Space: 43.2 MB (37.38%) - LeetHub

+++ 0389-find-the-difference/0389-find-the-difference.java
@@ -9,7 +9,7 @@ class Solution {
         }
 
         for(int i=0;i<m.length;i++){
-            xor=xor^ m[i];
+            xor=xor^(int) m[i];
         }
